fix: upgrade go-httpauth to support basic auth
Merge and add proxy and proxy auth related tests
