+++ cliv2/go.mod
@@ -5,7 +5,7 @@ go 1.18
 require (
 	github.com/elazarl/goproxy v0.0.0-20220328115640-894aeddb713e
 	github.com/hashicorp/go-uuid v1.0.3
-	github.com/snyk/go-httpauth v0.0.0-20220912133144-d51219ca664d
+	github.com/snyk/go-httpauth v0.0.0-20220915135832-0edf62cf8cdd
 	github.com/spf13/viper v1.13.0
 	github.com/stretchr/testify v1.8.0
 )
