+++ test/fixtures/squid_environment/Dockerfile
