+++ test/fixtures/squid_environment/docker-compose.yml
