+++ test/fixtures/squid_environment/scripts/setup.sh
